fix: jpa open-in-view false 적용

+++ src/main/java/kr/co/pinpick/archive/service/ArchiveService.java
@@ -10,11 +10,15 @@ import kr.co.pinpick.archive.repository.ArchiveLikeRepository;
 import kr.co.pinpick.archive.repository.archive.ArchiveRepository;
 import kr.co.pinpick.archive.repository.ArchiveTagRepository;
 import kr.co.pinpick.common.dto.PaginateResponse;
+import kr.co.pinpick.common.error.BusinessException;
+import kr.co.pinpick.common.error.ErrorCode;
 import kr.co.pinpick.user.dto.response.FolderDetailResponse;
 import kr.co.pinpick.user.entity.Folder;
 import kr.co.pinpick.user.entity.User;
 import kr.co.pinpick.user.repository.FolderArchiveRepository;
+import kr.co.pinpick.user.repository.FolderRepository;
 import kr.co.pinpick.user.repository.FollowerRepository;
+import kr.co.pinpick.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
@@ -37,7 +41,9 @@ public class ArchiveService {
     private final ArchiveLikeRepository archiveLikeRepository;
     private final FollowerRepository followerRepository;
     private final GeometryFactory geometryFactory;
+    private final FolderRepository folderRepository;
     private final FolderArchiveRepository folderArchiveRepository;
+    private final UserRepository userRepository;
 
     @Transactional
     public ArchiveResponse create(User author, CreateArchiveRequest request, List<MultipartFile> attaches) {
@@ -90,8 +96,8 @@ public class ArchiveService {
             archiveIds.add(a.getId());
         });
 
-        Map<Long, Boolean> isFollowMap = getIsFollowMap(user, authorIds);
-        Map<Long, Boolean> isLikeMap = getIsLikeMap(user, archiveIds);
+        var isFollowMap = getIsFollowMap(user, authorIds);
+        var isLikeMap = getIsLikeMap(user, archiveIds);
 
         return ArchiveCollectResponse.builder()
                 .collect(archives
@@ -113,14 +119,19 @@ public class ArchiveService {
     }
 
     @Transactional(readOnly = true)
-    public ArchiveResponse get(User user, Archive archive) {
+    public ArchiveResponse get(User user, Long archiveId) {
+        var archive = archiveRepository.findByIdOrElseThrow(archiveId);
+        if (!archive.getIsPublic() && !archive.getAuthor().getId().equals(user.getId())) {
+            throw new BusinessException(ErrorCode.UNAUTHORIZED);
+        }
         var isFollow = followerRepository.existsByFollowerAndFollow(user, archive.getAuthor());
         var isLike = archiveLikeRepository.existsByAuthorAndArchive(user, archive);
         return ArchiveResponse.fromEntity(archive, isFollow, isLike);
     }
 
     @Transactional(readOnly = true)
-    public ArchiveCollectResponse getByUser(User user, User author) {
+    public ArchiveCollectResponse getByUser(User user, Long authorId) {
+        var author = userRepository.findByIdOrElseThrow(authorId);
         var isAuthor = user.getId().equals(author.getId());
         List<Archive> archives;
         if (isAuthor) {
@@ -129,7 +140,7 @@ public class ArchiveService {
             archives = archiveRepository.findAllByAuthorAndIsPublic(author, true);
         }
 
-        Set<Long> archiveIds = archives.stream().map(Archive::getId).collect(toSet());
+        var archiveIds = archives.stream().map(Archive::getId).collect(toSet());
         var isFollow = followerRepository.existsByFollowerAndFollow(user, author);
         var isLikeMap = getIsLikeMap(user, archiveIds);
         return ArchiveCollectResponse.builder()
@@ -139,7 +150,8 @@ public class ArchiveService {
     }
 
     @Transactional(readOnly = true)
-    public ArchiveCollectResponse getByFolder(User user, Folder folder) {
+    public ArchiveCollectResponse getByFolder(User user, Long folderId) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
         var folderArchives = folderArchiveRepository.getByFolder(folder);
         var archiveIds = folderArchives.stream().map(fa -> fa.getArchive().getId()).collect(toSet());
         var isLikeMap = getIsLikeMap(user, archiveIds);
@@ -151,20 +163,31 @@ public class ArchiveService {
     }
 
     @Transactional(readOnly = true)
-    public FolderDetailResponse getArchivesWithFolderInfo(User user, Folder folder) {
+    public FolderDetailResponse getArchivesWithFolderInfo(User user, Long folderId) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
         var archiveIds = folder.getFolderArchives().stream().map(fa -> fa.getArchive().getId()).collect(toSet());
         var isLikeMap = getIsLikeMap(user, archiveIds);
         return FolderDetailResponse.fromEntity(folder, user, isLikeMap);
     }
 
     @Transactional
-    public void delete(Archive archive) {
+    public void delete(User author, Long archiveId) {
+        var archive = archiveRepository.findByIdOrElseThrow(archiveId);
+        checkAuthorization(author, archive);
         archiveRepository.delete(archive);
     }
 
     @Transactional
-    public Boolean changeIsPublic(Archive archive, boolean isPublic) {
+    public Boolean changeIsPublic(User author, Long archiveId, boolean isPublic) {
+        var archive = archiveRepository.findByIdOrElseThrow(archiveId);
+        checkAuthorization(author, archive);
         archive.setIsPublic(isPublic);
         return archive.getIsPublic();
+    }
+
+    private void checkAuthorization(User author, Archive archive) {
+        if (!archive.getAuthor().getId().equals(author.getId())) {
+            throw new BusinessException(ErrorCode.UNAUTHORIZED);
+        }
     }
 }
