+++ src/main/java/kr/co/pinpick/user/service/FolderService.java
@@ -1,6 +1,7 @@
 package kr.co.pinpick.user.service;
 
 import kr.co.pinpick.archive.entity.Archive;
+import kr.co.pinpick.archive.repository.archive.ArchiveRepository;
 import kr.co.pinpick.common.dto.PaginateResponse;
 import kr.co.pinpick.common.error.BusinessException;
 import kr.co.pinpick.common.error.ErrorCode;
@@ -13,6 +14,7 @@ import kr.co.pinpick.user.entity.User;
 import kr.co.pinpick.user.repository.FolderArchiveRepository;
 import kr.co.pinpick.user.repository.FolderRepository;
 import kr.co.pinpick.user.repository.FollowerRepository;
+import kr.co.pinpick.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,9 +25,11 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class FolderService {
+    private final UserRepository userRepository;
     private final FolderRepository folderRepository;
     private final FolderArchiveRepository folderArchiveRepository;
     private final FollowerRepository followerRepository;
+    private final ArchiveRepository archiveRepository;
 
     @Transactional
     public FolderResponse create(User user, CreateFolderRequest request, MultipartFile attach) {
@@ -41,7 +45,8 @@ public class FolderService {
     }
 
     @Transactional(readOnly = true)
-    public FolderCollectResponse getFolderList(User user, User author) {
+    public FolderCollectResponse getFolderList(User user, Long authorId) {
+        var author = userRepository.findByIdOrElseThrow(authorId);
         var isAuthor = user.getId().equals(author.getId());
         List<Folder> folders;
         if (isAuthor) {
@@ -56,7 +61,9 @@ public class FolderService {
     }
 
     @Transactional
-    public void addArchiveToFolder(User user, Folder folder, Archive archive) {
+    public void addArchiveToFolder(User user, Long folderId, Long archiveId) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
+        var archive = archiveRepository.findByIdOrElseThrow(archiveId);
         var isFollow = followerRepository.existsByFollowerAndFollow(user, archive.getAuthor());
         if (!isFollow && !user.getId().equals(archive.getAuthor().getId())) {
             throw new BusinessException(ErrorCode.ONLY_ADDABLE_FOLLOWERS_ARCHIVE);
@@ -69,7 +76,9 @@ public class FolderService {
     }
 
     @Transactional
-    public void removeArchiveFromFolder(Folder folder, Archive archive) {
+    public void removeArchiveFromFolder(Long folderId, Long archiveId) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
+        var archive = archiveRepository.findByIdOrElseThrow(archiveId);
         var folderArchiveOptional = folderArchiveRepository.findByFolderAndArchive(folder, archive);
         if (folderArchiveOptional.isEmpty()) {
             throw new BusinessException(ErrorCode.FOLDERARCHIVE_ALREADY_REMOVED);
@@ -78,13 +87,23 @@ public class FolderService {
     }
 
     @Transactional
-    public Boolean changeIsPublic(Folder folder, boolean isPublic) {
+    public Boolean changeIsPublic(User user, Long folderId, boolean isPublic) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
+        checkAuthorization(user, folder);
         folder.setPublic(isPublic);
         return folder.isPublic();
     }
 
     @Transactional
-    public void deleteFolder(Folder folder) {
+    public void delete(User user, Long folderId) {
+        var folder = folderRepository.findByIdOrElseThrow(folderId);
+        checkAuthorization(user, folder);
         folderRepository.delete(folder);
+    }
+
+    private void checkAuthorization(User user, Folder folder) {
+        if (!folder.getUser().getId().equals(user.getId())) {
+            throw new BusinessException(ErrorCode.UNAUTHORIZED);
+        }
     }
 }
